refactor : 파일 분할
from
/server/server.c

to
/server/include/server.h
/server/include/connection.h
/server/include/signal_handlers.h
/server/include/utils.h
/server/src/server.c
/server/src/connection.c
/server/src/signal_handlers.c
/server/src/utils.c
