+++ CMakeLists.txt
@@ -13,7 +13,11 @@ set(CLIENT_SOURCES
 )
 
 set(SERVER_SOURCES
-    ${CMAKE_SOURCE_DIR}/server/server.c
+    ${CMAKE_SOURCE_DIR}/server/main.c
+    ${CMAKE_SOURCE_DIR}/server/src/server.c
+    ${CMAKE_SOURCE_DIR}/server/src/connection.c
+    ${CMAKE_SOURCE_DIR}/server/src/signal_handlers.c
+    ${CMAKE_SOURCE_DIR}/server/src/utils.c
 )
 
 add_executable(client ${CLIENT_SOURCES})
