+++ server/include/signal_handlers.h
@@ -1,0 +1,9 @@
+#ifndef SIGNAL_HANDLERS_H 
+#define SIGNAL_HANDLERS_H 
+
+void setup_signal_handlers();
+void child_zombie_handler(int signo);
+void sigusr1_handler(int signo);
+void sigusr2_handler(int signo);
+
+#endif
