+++ server/main.c
@@ -1,0 +1,9 @@
+#include "server.h"
+#include <stdio.h>
+#include <stdlib.h>
+
+int main(int argc, char** argv) {
+    init_server();
+    run_server();
+    return 0;
+}
