+++ server/src/utils.c
@@ -1,0 +1,7 @@
+#include "utils.h"
+#include <fcntl.h>
+
+void set_nonblock(int fd) {
+    int flags = fcntl(fd, F_GETFL);
+    fcntl(fd, F_SETFL, flags | O_NONBLOCK);
+}
